Refactor the inserter menu component and split into multiple smaller components
